protect mcp.json from deletion

+++ _playground/clear-playground.sh
@@ -7,8 +7,18 @@ cd vscode-app-folder
 find . -mindepth 1 -not -name '.gitignore' -exec rm -rf {} +
 
 # if mcp.json exists
-if [ -f ~/.aws/amazonq/mcp.json ]; then
-    cp ~/.aws/amazonq/mcp.json ~/.aws/amazonq/mcp.json.$(date +%Y%m%d_%H:%M:%S)
+if [ -f ~/.aws/amazonq/mcp.json -a "$1" == "--mcp-server" ]; then
+
+    save_name="~/.aws/amazonq/mcp.json.$(date +%Y-%m-%d-%H-%M-%S)"
+    echo ""
+    echo "~/.aws/amazonq/mcp.json exists and will be replaced"
+    read -p "Do you want to save it as $save_name? (y/n) " answer
+
+    if [ "$answer" != "${answer#[Yy]}" ] ;then
+        echo "Saving file $save_name"
+        cp ~/.aws/amazonq/mcp.json $save_name
+    fi
+
     rm ~/.aws/amazonq/mcp.json
 fi
